refactor: move ui logic to index.tsx
